Initial check-in of nemo-rabbitvcs modules.
This is a port of the rabbitvcs-nautilus module to nemo. It is a
straight name-swap from nautilus to git, with two notable changes.

1. RabbitVCS expects to find the environment variable
NAUTILUS_PYTHON_REQUIRE_GTK3 set if it should load GTK3. The name of
this variable cannot be changed as a result.

2. make_menu_item() constructs a proper Nemo menu item in the extension
code. The original extension called back into the rabbitvcs-core code,
but there was no reason for it to do so.

(cherry picked from commit e175feb58f37dc768636862e62278f6af1e9b188)